Repackage + Builder improvements
