+++ textile-recipe-api-v1/src/main/java/net/textilemc/textile/api/recipe/v1/SimpleDispatcher.java
@@ -1,4 +1,4 @@
-package net.textilemc.textile.api.recipe.event.v1;
+package net.textilemc.textile.api.recipe.v1;
 
 import java.util.ArrayList;
 import java.util.List;
