+++ textile-recipe-api-v1/src/main/java/net/textilemc/textile/impl/recipe/DispatcherHolder.java
@@ -1,0 +1,9 @@
+package net.textilemc.textile.impl.recipe;
+
+import java.util.LinkedList;
+
+import net.textilemc.textile.api.recipe.v1.Dispatcher;
+
+public class DispatcherHolder {
+	public static final LinkedList<Dispatcher<?>> DISPATCHERS = new LinkedList<>();
+}
