+++ textile-recipe-api-v1/src/main/java/net/textilemc/textile/mixin/recipe/RecipeDispatcherMixin.java
@@ -21,7 +21,7 @@ import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 
-import net.textilemc.textile.api.recipe.event.v1.ShapedRecipeCallback;
+import net.textilemc.textile.api.recipe.v1.event.ShapedRecipeCallback;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
